requirements.txt
streamlit==1.28.1
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
accelerate==0.24.1
sentencepiece==0.1.99
protobuf==4.24.4
llava-next
pandas>=1.5.0 